Restore missing OP_PUSHMARK before our arguments.
An entersub has variable arguments, so usually the optree will
look like this:

    entersub
        pushmark
            argop1
            argop...
            argopN
        gvcv

For a custom opcode that enforces the number of arguments at compile time
(that is, it's enforcing a prototype on an opcode) the pushmark is
unnecessary, because you just pop from the stack without checking.

For a custom opcode that does NOT enforce a prototype (either because
it has variable arguments, or because it wants to allow passing a list,
ala foo(qw/1 2 3/)), the pushmark is necessary.

With that pushmark, TOPMARK will be whatever it was when the
subroutine was invoked.  If whatever invoked the sub was an xsub, then
the state of the stack at the point where the xsub was invoked is
what will be used, plus/minus whatever stack manipulations happened
in the meanwhile.

*Ideally* we would reuse the entersub pushmark, but I keep failing
at splicing the optree properly, so instead this just allocates a new
OP_PUSHMARK.